Update patch set 12
Patch Set 12:

> Build Failed
 > 
 > https://jenkins.opendaylight.org/releng/job/netvirt-patch-test-current-neon/285/
 > : FAILURE (skipped)
 > 
 > The job has timed out after 360 minutes. Examine output, retry, or
 > consider increasing timeout value for this job type/stream. (
 > https://jenkins.opendaylight.org/releng/job/netvirt-patch-test-current-neon/285/
 > )
 > 
 > Logs: https://logs.opendaylight.org/releng/vex-yul-odl-jenkins-1/netvirt-patch-test-current-neon/285

The csit job did finish, the problem is the test-patch parent job started and triggers the child csit job - but that didn't start for a few hours. When the csit did run, it ran for 4 hours and by then the total aggregate job time was now over the 6 hour limit. Below is the link to the child csit job. only way to correlate it to the parent job, is to start with the link below without the job number at the end, then click on each job and look for the gerrit listed to match the gerrit you have.

https://jenkins.opendaylight.org/releng/job/netvirt-csit-1node-0cmb-1ctl-2cmp-openstack-queens-gate-stateful-neon/400/

Patch-set: 12
Label: Verified=0