neutron-vpn-portip-port-data d.s not cleared
Issue:
======
Stale entry not removed for neutron-vpn-portip-port-data in NAT
use-case.

neutron-vpn-portip-port-data d.s has ipv4 ip's which is
not getting cleared after  all the configs are deleted.

Analysis:
=========
The root-cause is in case of NAT use-case when external gateway is
attached to  router, while looping through the list of ip's addition
of ipv4 ip's to vpn(internetbgpvpn) was not restricted but while
dissasociating the vpn this restriction   check is there.
        So ipv4 ip's were retained in data-store.
Added check in updateVpnInterface() in NeutronvpnManager.java
to not add ipv4 entries in neutron-vpn-portip-port-data in case of
internetbgvpn (passed the flag isInternetVpn) where isInternetVpn
flag will be true in this case.

Issue: NETVIRT-1406

Change-Id: I9ae2b1da04988aac461ff50c4a9ac5a9d80a370d
Signed-off-by: nithi thomas <nithi.t@altencalsoftlabs.com>