Migrate to genius/vniutils
JIRA: GENIUS-250

Netvirt currently has a vni carving logic which is
tightly coupled with NAT. There are more netvirt
services which are in need of this utility now, and
hence we need to move this to a centralized place.
Adding a new VniUtils in genius, to which now nat is getting migrated to.

Change-Id: Ia08ad6eba9a6b5000c59dd804c8e50f17907257a
Signed-off-by: Faseela K <faseela.k@ericsson.com>