Update patch set 6
Patch Set 6:

> Sam, this seems to always fail those test on something. Last time
 > it was a 6 hour timeout.
 > 
 > How do you propose we go about this?

This last run timed out because of the infra load. The actual csit did eventually run but it ran much later than the parent job. The root job started and builds odl, then that triggers the csit to run on that distro. But that child csit job didn't start for 3 hours. Csit takes 3-4 hours, so the total time is more than 6 hours and that is the timeout that the root job reports.

The csit that did run is at the link below. It had one failure.

https://jenkins.opendaylight.org/releng/job/netvirt-csit-1node-0cmb-1ctl-2cmp-openstack-queens-gate-stateful-neon/463/robot/

Patch-set: 6
Label: Verified=0