auto-config-transport-zones has incorrect default
Fix for NETVIRT-1009 modified the default in config
file for auto-config-transport-zones to false. But
it missed changing in yang file. When a param is
missing in config file, it picks off yang defaults.
This caused a discrepency in behavior due to conflicting
defaults between two.

This fixes yang default to reflect the config default
as per NETVIRT-1009.

This also fixes the log message for two listeners which
used this flag and were misleading.

Change-Id: Ic1b85b798eedd6a14a80eb6145b5a23f73923fdc
JIRA: NETVIRT-1546
Signed-off-by: Vishal Thapar <vthapar@redhat.com>