Issue: Table 51 MAC issue after CSC reboot
Description: If MAC movement happens after a CSC reboot,
the old entry in CSC gets cleared followed by a idle timeout.
However, this pushes a flow delete to CSS which removes
the working flow as well due to same match condition.

The solution is to construct a flow key with ELAN tag and MAC address,
which would update the same flow after MAC movement.

Change-Id: Icad172d2c446cbf8bb2e6d12a8a8117b02057925
Signed-off-by: eswanit <swati.udhavrao.niture@ericsson.com>