Update patch set 10
Patch Set 10: Code-Review-1

(5 comments)

sorry but logging is wrong here, can you fix as suggested?  Also, it would be better to "propagate" (i.e. just add "throws ReadFailedException" to those those, and let the caller "send it up"), but that is sometimes a bit more work, so you can do it in 2 steps and first do this and then follow-up with another change which removes the catch and LOG and propagates it...

Patch-set: 10
Label: Code-Review=-1