Dual Stack feature support for router interface
This is second patch to support dual stack
router interface

We create a dual-stack neutron port and attach this
port to the router.
This single neutron port is then responsible for
handling both IPv4 and IPv6 subnets in that network.
The advantage of this approach is that it reduces
the number of Neutron router ports.

Issue: NETVIRT-1543

Change-Id: Iabd96ec2ad60ee8e5a3078f25304c175117830d0
Signed-off-by: Karthikeyan Krishnan <karthikeyangceb007@gmail.com>