Update patch set 11
Patch Set 11:

Pasting Dhiman's comments here:

String.intern() is just a bad idea while using it as monitor lock in synchronization.
1.	If we google around, can see how many places people have faced unwanted deadlocks.
2.	From Java 7 the String Pool is part of the heap moved from perm gen, so when the interned string will be GCed we never know, it could lead to synchronization failures.

I think pooled ReenterentLocks is good idea as long as we have a strategy when to remove a lock from the pool. What strategy is being used in JvmGlobalLocks class?

I agree try/finally is few more lines of code and we must carefully identify the code path where to call the lock and unlock, but considering the risk of using interned string, it’s better.

Patch-set: 11
Label: Code-Review=0