Use a single transaction for removeMacsFromElanExternalDevices()
removeMacsFromElanExternalDevices() processes multiple devices
and macs and calls out to HwvtepUtils, which allocates and submits
a single transaction for each request. Furthermore the list of
macs is translated multiple times if there are multiple devices
involved.

This patch refactors this method to lazily allocate both the
transformed list and a write transaction, committing all changes
in one go, rather than in multiple transactions.

Change-Id: Ie15e9feea76a3820479b2836d81f7ac0e7c9d9fb
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>