Update patch set 1
Patch Set 1: Code-Review+1

And this makes it much clearer what needs to be done: we really want to make sure to reuse whatever transaction context we are in to encapsulate operations.

This really means that the various Utils which are allocating transactions are deprecated and should be eliminated as soon as possible, so that transactions are allocated only at high-order methods.

Patch-set: 1
Reviewer: Gerrit User 649 <649@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1