MIP related improvement
List of changes done to fix issues related to MIP.

1) LearntVpnVipToPort yang container is removed. MIPs were stored in this
operational DS. But when cluster-reboot is performed, this container will
not have any data.To start the aliveness-monitoring sessions post
cluster-reboot,we needed a container to store the MIP which can be
persisted. Hence MIPs are now being stored in NeutronVpnPortIpToPort
config DS.

2) For a pre-created port, when VM is deleted ,
InterfaceStateChangeListener remove event is received. If this VM hosts a
MIP, then MIP adj needs to be removed from Cfg-Vpn-Interface DS.

Change-Id: I144f180961d4bd4c0829571d132c1d706f7892de
Signed-off-by: HANAMANTAGOUD V Kandagal <hanamantagoud.v.kandagal@ericsson.com>