Update patch set 2
Patch Set 2:

> Please dont remove the use of Resource-Batching-Manager(RBM) for
 > BGP routes. We want to know the motive behind removing the use of
 > it.

We’ve been discussing this for the past couple of years. The motivation behind removing the use of batched transactions is that they break the definition of transactions: one transaction is one set of related operations, which should be persisted as a unit or not at all. The approach we talked about pursuing, back in 2017 IIRC, was to make the transactions larger, which is what this PR does, instead of relying on a transaction batcher.

Patch-set: 2