Close All MDSAL read transcations
When a MIP is learned via a VPNInterface,
we create a secondary adjacency (carrying the MIP)
for that VPNInterface. However on a cluster reboot
(or node reboot), since we do a direct txn.put()
on the VPNInterface, the MIP adjacency permanently
disappears. This in-turn triggers the VPNEngine to
remove the FIBEntries and flows for the MIP, ending
up with a dataplane loss for the MIP.

After this situation, even if user tries to ping the
MIP, since the MIP is still contained in
NeutronVpnPortIpToPort, the ArpReponse (or GARPRequests)
packets carrying the MIP are ignored. So all in all
, we end up with permanent dataplane loss for MIP
on a cluster reboot / node reboot. This fix attempts
to retain the MIP secondary adjacency on the
VPNInterface (instead of inadvertantly removing it)
thereby preventing the MIP/FIB disappearance and
providing us lossless dataplane for MIPs,
on either cluster/config-shard -leader-nodal reboots.

TR: HX53335 - N9 BigLab: Always Close all the MDSAL
Read Transactions after use Transaction has not been
closed for ReadOnlyTransaction which leads to
unnecessary delay in destroying the shard
traction actors (in case of remote reads). This would
lead to memory leak and can also cause delay in creating
data tree snapshots for reads/writes. As a solution
tx.close() is been added to all ReadOnlyTransaction.

Change-Id: I3977f708ae3427d2f371c228d6ba4d773e9d74cd
Signed-off-by: Surendar Raju <surendar.raju@ericsson.com>