VRF not written by VPNEngine into ebgp datastore
This fix is crucial in addressing three important issues:

a. There was no serialization between add() method
and update() method within the VpnInstanceListener
and this was resulting in reordering of events.
In this we fix update() method such that it
also uses the queuing to order the event
processing similar to add() and update().

b. This fix also drives isVpnInstanceReady() notification
at the right moment for use by all waiters on a
VPN-Creation, which is the moment when the
ADDRESS-FAMILY is configured for that VPN.

c. There was also no validation when a VPN is
created via Neutron BGPVPN, if such a VPN already
exists in the CSC System operational-DS as
vpn-instance-op-data.
This fix includes a validation where it will
disallow configuring a new VPN if another VPN
with the same RD exists in our
operational-ds vpn-instance-op-data.

Change-Id: I8e6f270990b9773835912bd4d61ed30cb5f14d7e
Signed-off-by: Surendar Raju <surendar.raju@ericsson.com>