Subnet route is not advertised to FIB table...
Issue:In this there is vm present in vpn.While
deleting this type of vpn, the entire vrf entry
is removed before the actual vpn is deleted.
Hence transcation commit failure happens,
which results in breakoff inbetween.

Solution:This scenario is expected, but to
avoid the break-off we had added fix to
handle the exception, which results in flow
continuous.

Change-Id: I959cfc12d3490210235cf853d03fe2562c27be47
Signed-off-by: Surendar Raju <surendar.raju@ericsson.com>