NAPT SwitchOver not happening properly
Description: When the NAPT switch was brought down, it is observed that
the other non-napt switch was not elected as NAPT

Root cause : The neutron-router-dpns DS had an entry of the old NAPT
switch that was brough down, the entry should have been delted when
the node was brought down. This prevented the NAT listener to elect
a new NAPT switch

Solution : Wrote a scheduler job that is trigerred when the node was
brought down and checks if there is any stale entry in the
neutron-router-dpns and deletes it.

ISSUE : NETVIRT-1614

Change-Id: I6bed9466de93d7c231e29c3f12eec305d1c49fb9
Signed-off-by: Chetan Arakere Gowdru <chetan.arakere@altencalsoftlabs.com>