IPv6 Stale enties in the compute which is evacuated.
Multiple issues fixed in this review.

Issue-1

Change-Id: I0697e24dce1e2990a6c703a66716209a203e751c
--------
Below are the 3 sets of flows per IPv6 subnet/network which is
left out on switch even when there are no VM's belonging to the
corresponding subnet/network on the switch. These flows used to gets
deleted when subnet is removed from the router.
a. Icmpv6NsDefaultPuntFlow
b. Icmpv6RSPuntFlow
c. Icmpv6NaPuntFlow

Reason for (a):
While handling InterfaceState remove event,
vnet.updateDpnPortInfo() was updated before removing the
corresponding flows, hence the check for last VM on
dpn was failing as the count was zero.

Reason for (b) and (c):
Removing these flows was not handled when it is the last VM on the dpn.

Solution:
---------
+ During InterfaceState remove event, vnet.updateDpnPortInfo() is updated
  after removing the corresponding flows.
+ For (b) and (c), method programIcmpv6PuntFlows() is updated to handle
  both add and delete actions. Thereby removes the corresponding flows
  during InterfaceState.remove event.

Issue-2:
--------
Race condition while updating VirtualNetwork.dpnIfaceList.ofPortMap object
during InterfaceState add/remove events.

Solution:
---------
During InterfaceState add event, ifMgr.updateDpnInfo() was called outside
DJC which is keyed on portId. Whereas vnet.updateDpnPortInfo was called
within DJC on the same key. This was resulting in race condition when
there are back-to-back add/remove events. This has been corrected.
Add/remove to DpnInterfaceInfo.ofPortMap is handeled within DJC on same
key during InterfaceState add/remove event.

Issue-3:
--------
Let's say two or more VM's are up on a dpn and later add the corresponding
subnet to router, Icmpv6NsDefaultPuntFlow are not programmed.
																								Solution:
---------
Corrected checkin ifMgr.checkIcmpv6NsMatchAndResponderFlow()
for handling router interface add.

Change-Id: I962aafe810e3413003c5ea222b5a96283f5b1980
Signed-off-by: Nishchya Gupta <nishchyag@altencalsoftlabs.com>