Update patch set 18
Patch Set 18:

(1 comment)

Vinod,

Thanks for addressing the changes. Do we need drop rules for remote SG as well. Like when we remove an SG with remote rule from a port we need to add drop for all removed rules.

Eg, SG1 -> Allow all ICMP

Vm1 default
Vm2 default SG1.

Ping vm2 to vm1 . Now remove default  from VM2. The traffic will continue as we are just removing the allow rule in case of vm1 but no drop rule. If we stop and start ping it will be blocked.

This can be handled in a separate patch as well.

Patch-set: 18
Label: Verified=0