Update patch set 2
Patch Set 2: Code-Review-1

(2 comments)

Vinod,
We have a aclFlowPriorityMap in AclDataUtil, where the key is the flow name. I think we can reuse the  data structure such that first part of the key represents the protocol and the port/icmp code and type. So if the first part of the key is present we can avoid adding the drop rule. Else we will be fixing the issue by introducing a new one. Please do share if you have any alternate thoughts.

Patch-set: 2
Label: Code-Review=-1