Update patch set 1
Patch Set 1:

We have tested the protocol number = 0 with openstack without ODL and found that all protocols (ICMP,TCP and UDP) communication are working well. So we have implemented in a way to support all the protocols for protocol number 0. So, before abandoning the patch, can we comment on this behaviour?

Patch-set: 1