Update patch set 1
Patch Set 1:

After setting the firewall driver as OVS firewall, I have retested the test case again and observed that the flows [1] [2] are installed for protocol number 0 and communication (ICMP, TCP an dUDP) is not happening  between VMs. So, protocol number 0 is not used as generic if we use OVS firewall.
 
 [1] cookie=0xa99821b3245d1aa6, duration=10.401s, table=82, n_packets=0, n_bytes=0, reset_counts priority=70,ct_state=+est-rel-rpl,ip,reg5=0x5,dl_dst=fa:16:3e:54:b8:be,nw_proto=0 actions=pop_vlan,output:5

 [2] cookie=0xa99821b3245d1aa6, duration=10.400s, table=82, n_packets=0, n_bytes=0, reset_counts priority=70,ct_state=+new-est,ip,reg5=0x5,dl_dst=fa:16:3e:54:b8:be,nw_proto=0 actions=ct(commit,zone=NXM_NX_REG6[0..15]),pop_vlan,output:5

Patch-set: 1