The entries for VPN-instances are not deleted after deleting the VPNs
This fix primarily address the presence of stale config vpn-instances
in the system even though such vpn-instances have been deleted either
by ODL RPC (or) Openstack BGPVPN.

In addition, we make sure to us the locking/unlocking correctly
whenever we are writing content to vpnMaps and vpnInstances
datastores of the config-DS.

A lock-acquire-wait-ceiling time of 10 seconds is used which is actually
large and even that wait-time need to be revisited.

It is unclear why we need protection only for these two datastores when
there are also other important datastores like subnetMap, networkMap etc
and so this needs to be reviewed.

Change-Id: I0ca3d7ac7753dac1c5c0c58b6fdfdd063366810f
Signed-off-by: Ashik Alias <ashik.alias@ericsson.com>