Update patch set 1
Patch Set 1:

There are two approaches to address this.
One is this approach in this patch.
The con of this approach is mdsal read and waiting.

The other is adding listeners as suggested by sam.
If this approach is taken, then some race conditions have to be addressed like, the vlan binding got deleted from config just about when the new logical switch listener is about to push the vlan binding.
Then we end up having stale vlan bindings in the device.

We could as well address this in plugin like the way ucast and mcast and logical switches are ordered.

Patch-set: 1