Update patch set 6
Patch Set 6:

Hi Poovizhi,

I'd vote to have the core logic in a separate class where you pass mdsalutil in the constructor. We should not be instatiating the utils class in the blueprint and it should be used to prvide getter/setter utilities only.
Adding Kiran to comment on it further.

Patch-set: 6