Update patch set 11
Patch Set 11: -Code-Review

Had given a +1 since in the original spec tox build, I had pointed out warnings to be fixed. Swati was looking to keep the fix limited to handling tox warnings but not correcting formatting as a whole.

Yes, there are quite a few formatting issues here..now if we are going to handle these here, as indicated by Vishal's -1, need to take back my upvote :)

Patch-set: 11