Update patch set 3
Patch Set 3:

Hi Hanamanta,

Let's do as we just discussed.
Having one for loop to loop over the interfaces, fetch the infName and do a syncWrite to adj identifier.

This will help us avoid:
1) two for loops
2) reading for infName twice
3) maintaining an unnecessary map
4) read on vpninf tree
5) perform a direct write instead of compare and update

Since we anyway are going to revisit this when we move to nexthop leaf, it's your wish to how much extent you want to fix it now :)

Patch-set: 3