Update patch set 1
Patch Set 1: Code-Review-1

(5 comments)

Hanamanta, some minor comments for the change but some design aspect which I request both you and Vivek to have a look at, hence a -1.

Vivek, please provide your opinion.
In the model below, acc to me, nexthop should be retained as a leaf element as was earlier.
    grouping adjacency-list{
        list adjacency{
            key "ip_address";
            leaf-list next-hop-ip-list { type string; }
            leaf ip_address {type string;}
            leaf primary-adjacency {
                type boolean;
                default false;
                description 
                    "Value of True indicates this is a primary adjacency";
            }
            leaf label { type uint32; config "false"; } /* optional */
            leaf mac_address {type string;} /* optional */
        }
    }

Patch-set: 1