Update patch set 16
Patch Set 16:

There is no need to manually add the wiring for neutronvpn as it has been converted to singleton injection. This holds for all the modules in netvirt exception elanmanager and vpnmanager which have not been converted yet.

The last patchset passed sft, but now the karaf 4 sft is failing but I don't think that is related to the patch.

Patch-set: 16
Label: Verified=0