Update patch set 2
Patch Set 2:

I'll review this and further comment on it when it passes the build...

If the test failure you're hitting is related to order of elements in a YANG system sorted List, then please see how we've "worked around" a similar problem in https://git.opendaylight.org/gerrit/#/c/60313/10/interfacemanager/interfacemanager-impl/src/test/java/org/opendaylight/genius/interfacemanager/test/InterfaceManagerConfigurationTest.java ... does looking at the help you here? It's a problem of Binding (v1; v2 is better) for which it would be cool to have a better generic solution, but we currently haven't be able to find one.

Patch-set: 2
Label: Verified=0