Update patch set 13
Patch Set 13: Code-Review-1

Earlier comment unanswered:
when we create an EVPN, the l2vni will not be available right?
It becomes available only when the first (and only) network is associated to that EVPN.. So why is this l2vni requested as an input for RPC here?

Patch-set: 13
Label: Code-Review=-1