Update patch set 4
Patch Set 4:

> (1 comment)

Actually, There will not be a case where prefix is null.
During advertise route i can add a check 
if (macentry.getIpPrefix== null) then return (without advertising)
But EVPN RT2, mac is a mandatory and IP is optional. IP will be used in arp suppression case. So i thought to advertise even if it is not present (though thr will not be a case where IP is not present).
Now if you feel this is not right way, i can change it to 
if (macentry.getIpPrefix== null) then return (without advertising) just like i am doing in withdraw route case.

Patch-set: 4