Update patch set 3
Patch Set 3: -Code-Review

In controller snat I see that metadata("metadata=0x30d46/0xfffffe") is not of the external subnet and hence the reverse traffic for NFIP to FIP is not hitting this rule where the metadata will be external n/w after FIP translation, I am missing something here?

Right, Curently external network metadata is used fo progamming remote FIB flow. for MPLSOverGRE cases, we are only matching based on tun_id(0x11171) instead of metadeta.

    cookie=0x8000003, duration=151.404s, table=21, n_packets=9, n_bytes=626, priority=42,ip,metadata=0x30d44/0xfffffe,nw_dst=172.160.0.3 actions=set_field:0x11171->tun_id,load:0x200->NXM_NX_REG6[],resubmit(,220)
	
	cookie=0x90186a4, duration=2563.825s, table=36, n_packets=9, n_bytes=626, priority=5,tun_id=0x11171 actions=goto_table:44

But, currently I see for vlan/flat use-cases, this tun-id is missing(in actions) in remote-FIB flow and as a result DNAT-to-SNAT UC will not work(to be investigated)

 cookie=0x8000003, duration=168.894s, table=21, n_packets=0, n_bytes=0, priority=42,ip,metadata=0x30d46/0xfffffe,nw_dst=10.10.10.7 actions=load:0x700->NXM_NX_REG6[],resubmit(,220)
 
 cookie=0x90186ad, duration=166.691s, table=36, n_packets=0, n_bytes=0, priority=5,tun_id=0x11171 actions=goto_table:44

Patch-set: 3