Update patch set 3
Patch Set 3: Code-Review+1

Hi All,

I tried to debug further, but I couldn't figure out when ExtRouter entity will be created with the empty mac-address with the available CSIT INFO logs.(around 240 TC been executed as part of tempest.api.network)

With Current implementation, I see ExtRouter entity is created  when neuton router is set with external gate-way and deleted when this is unset. So, didn't find use-case with ExtRouter created with an empty ext-gw mac-address.

We can go ahead with this check to prevent this Exception. 

Whenever, the Ext-router is updated with ext-gw mac-address, the router gate-way flow/arp responder flows will be created properly(as part of ExternalRoutersListener.update())

Thanks,
Chetan

Patch-set: 3