Update patch set 2
Patch Set 2:

Based on initial investigation, it looks like there is a delay in installing table=21 flows when previous TC – “Associate L3VPN To Routers” is executed.

When this test-cases is executed, I see table=21 flows for many VMs are missing(in Teardown logs – for 10.1.1.4 & 20.1.1.4)

When the Failed TC – “Verify L3VPN Datapath With Router Association” is executed, few of them added back(10.1.1.4) but when condition is checked, table=21 flow for 20.1.1.4 is still missing resulting in this TC failure.

https://jenkins.opendaylight.org/releng/job/netvirt-csit-1node-openstack-newton-nodl-v2-gate-stateful-carbon/329/robot/report/log.html#s1-s4-t15-k8

When the Teardown logs are taken for this failed TC, the table=21 flow for VM 20.1.1.4 is visible.

https://jenkins.opendaylight.org/releng/job/netvirt-csit-1node-openstack-newton-nodl-v2-gate-stateful-carbon/329/robot/report/log.html#s1-s4-t15-k9-k1-k3-k1-k16-k4

cookie=0x8000003, duration=67.103s, table=21, n_packets=0, n_bytes=0, priority=42,ip,metadata=0x222e0/0xfffffffe,nw_dst=20.1.1.4 actions=group:150003

Patch-set: 2