Update patch set 1
Patch Set 1:

Hi Vivek,

I compared the size of the karaf logs with this change and others. I don't see any drastic change in size of karaf logs with this change.

With this fix.
[ ]	odl1_exception.log.gz	2017-04-11 05:38 	7.3K	 
[ ]	odl1_karaf.log.gz	2017-04-11 05:38 	1.3M	 
[ ]	odl1_karaf_console.log.gz	2017-04-11 05:38 	531 	

Other CSIT run
[ ]	odl1_exception.log.gz	2017-04-11 01:10 	6.3K	 
[ ]	odl1_karaf.log.gz	2017-04-11 01:10 	1.3M	 
[ ]	odl1_karaf_console.log.gz	2017-04-11 01:10 	531

Request to let me know if we can go ahead with this change.

Patch-set: 1