Update patch set 2
Patch Set 2:

Ocata CSIT job is having multiple failures to be addressed inside the CSIT scripts themselves.

So let us not wait for that to stabilize to land this patch.

cherrypick the patch to stable/carbon.

BTW, is SNAT working fine for both TCP/UDP with Ocata?

VIvek

Patch-set: 2
Label: Code-Review=+2