Update patch set 3
Patch Set 3:

Hi Vivek,

As I mentioned in the previous comment for this review. I'm adding this as dependent to https://git.opendaylight.org/gerrit/#/c/52996/.
And this will reduce the number of lines to minimal.
Please see that I had already given -1 for this.

Thanks,
Manu

Patch-set: 3
Label: Code-Review=-1