Update patch set 6
Patch Set 6:

Hi Vivek,

Do we have plans to fix it differently for Carbon which won't affect performance? Back then I had a thought with Tali on maybe using the "Listener Dependency Helper" in order to solve this.

Anyway, those problems existed in Boron for sure. If you still prefer to keep it out until we see it again, then I vote for the error you suggested - At least it will be easier to tell when it does happen

Patch-set: 6
Label: Verified=0