Update patch set 1
Patch Set 1:

1)
Can you elaborate on the problem a bit more?:)
Why doesn't the elanTag already exist in the elanInstance? Wasn't this data persisted along with the elanInstance itself in the configuration datastore?

2) 
Is this related only to cluster reboot? Or would it happen in a single-instance deployment reboot to? And if not, what is the difference?

Patch-set: 1