Update patch set 1
Patch Set 1:

Looking at "neutron subnet-update" in https://docs.openstack.org/cli-reference/neutron.html, I see that a subnet can never be updated with networkId or tenantId via OpenStack.
It was put in place when in vpnservice, majority of testing was via direct REST based APIs. Now that we've moved to OS for such orchestrations, doesn't make sense to have this in place.
Also, it simplifies updateSubnetNode to have 3 params, instead of 6 before this review.
This change is done in PS-2.

Patch-set: 1
Label: Verified=0