Update patch set 1
Patch Set 1:

Hi Guy,

Please find my responses below:

1)
Can you elaborate on the problem a bit more?:)
Why doesn't the elanTag already exist in the elanInstance? Wasn't this data persisted along with the elanInstance itself in the configuration datastore?

>> Before reboot, elanTag existed in the elan instance, but after reboot, following sequence of events take place:

Upon network ADD DCN, ElanInstance gets created again (replacing the previous one), then we listen on ELANInstance DS to get an ELANInstance ADD DCN (in ELANInstanceManager), following which IDManager allots a unique tag and updates the ELANInstance entity.

This process takes a longggg time and is okay the first time configs are done before a reboot since a network create is (generally) followed by:
1.       subnet create – time taken t1
2.       port create – time taken t2
3.       operations upon port add until ELANInstance is read – time taken t3

This races with:
4.       ELANInstance creation – time taken t4
5.       DCN fired, ID fetched, appended to ELANInstance entity – time taken t5

Before a reboot, in the normal case, always, (t1 + t2 + t3) > (t4 + t5) so we don’t see issues.
Upon reboot, with events fired closely, we’re seeing this race issue.

2) 
Is this related only to cluster reboot? Or would it happen in a single-instance deployment reboot to? And if not, what is the difference?

>> It would happen in single-instance deployment reboot too. No such difference.

Patch-set: 1