Update patch set 6
Patch Set 6:

Abhinav, Vivek - I think this change also fixes [1] which is a bug we saw happening several times in Boron. We pretty much forgot about it - and one of the possible fixes back then was to syncWrite the Oper transaction in VpnInterfaceManager (as done in this change).

Back then we thought it might have bad effects on performance, but if this change is in Carbon anyway - I think we should merge it (at least this part).

WDYT? 

[1] https://bugs.opendaylight.org/show_bug.cgi?id=7685

Patch-set: 6