Update patch set 1
Patch Set 1: Code-Review-1

> > the last csit run showed that same log. Does that mean the issue
 > is
 > > not fixed with this patch?
 > 
 > Hi Sam, this patch doesn't fix the issue, rather places an error
 > log to make the troubleshooting easier.
 > Rest all other changes are to avoid checkstyle failures..

Hi Sam,

I could not see the same issue in Carbon happening in Boron CSIT run for this patch as in [1].  And so having the error statement alone in this cherrypick is more than sufficient.

All the 26 occurrences of the line 
"Local DPNID is empty for rd" is for RouterInterface VRFEntries only and not for normal VRFEntries like in Carbon.

[1] https://logs.opendaylight.org/releng/jenkins092/netvirt-csit-1node-openstack-newton-nodl-v2-gate-stateful-boron/145/archives/odl1_karaf.log.gz

I will update this cherrypick to cover proper handling of RouterInterface VRF Entries too.

Vivek

Patch-set: 1