Update patch set 1
Patch Set 1:

Hi Abhinav/Vivek

Looking both on this review and https://git.opendaylight.org/gerrit/#/c/45998/ it seems that after cluster reboot there is an attempt to handle ordering issues while DS loads on a case-to-case basis.
I was wondering if there is or should be a plan for more robust cluster recovery implementation for Carbon?

Patch-set: 1
Label: Verified=0